pyproject.toml
[tool.poetry]
name = "marlenn"
version = "0.1.0"
description = ""
authors = ["Lario dos Santos Diniz <lariodiniz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
cryptography = "^39.0.2"
python-decouple = "^3.8"
django-extensions = "^3.2.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.10.3"
pip-audit = "^2.5.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.3"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
format = "blue . && isort ."
lint = "blue --check --diff . && isort --check --diff . && pip-audit"
docs = "mkdocs serve"
migrations = "python manage.py makemigrations && python manage.py migrate "
runserver = "python manage.py runserver 0.0.0.0:8000"
pre_test = "task lint"
test = "python manage.py runserver test"
